treewide: required yocto updates to UNPACKDIR usage
Yocto seems to have once again tightened the strings on how to correctly
use S, UNPACKDIR, and WORKDIR. These are the needed changes to get it to
compile.

There are migration instructions in migration-5.3.rst.

Also remove the general setting of S within files if it's only setting
it to a default.

Change-Id: I184922c609efebe4434ed71e9b381fd40159242d
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>